Move stubble around
prep for extra headers, builder map

+++ src/engine/quaternion.cpp
@@ -6,7 +6,7 @@
 
 using std::ostream;
 
-Quaternion::Quaternion(double w_, double x_, double y_, double z_) : w(w_), x(x_), y(y_), z(z_) {}
+Quaternion::Quaternion(float w_, float x_, float y_, float z_) : w(w_), x(x_), y(y_), z(z_) {}
 
 Quaternion Quaternion::operator*(const Quaternion& q) const {
 	return {
