+++ src/engine/quaternion.h
@@ -10,7 +10,7 @@ public:
 	float w, x, y, z;
 	Quaternion(); // 0 rotation
 	Quaternion(float angle, const Position3d& axis);// Use position3d vector as axis
-	Quaternion(double w_, double x_, double y_, double z_);
+	Quaternion(float w_, float x_, float y_, float z_);
 
 	Quaternion operator*(const Quaternion& q) const;
 	Quaternion conjugate() const;
